Refactor bgp logic, and fix BGP Reset
* Moved config conversion logic to new file.
* Reordered functions for top-down reading order.
* Changed local_tests to clear BGP gNMI config between subtests for a
  modest speed-up.
