+++ bgp/BUILD
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "bgp",
     srcs = [
+        "config.go",
         "gobgp.go",
         "ocgobgp.go",
         "util.go",
